pyproject.toml
[tool.poetry]
name = "ai-researcher"
version = "0.1.0"
description = ""
authors = ["Samuel Lehotsky <lehotsky@atlas.sk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
openai = "^1.12.0"
langchain = "^0.1.9"
python-dotenv = "^1.0.0"
unstructured = {extras = ["all-docs"], version = "^0.12.2"}
pypandoc-binary = "^1.12"
ebooklib = "^0.18"
beautifulsoup4 = "^4.12.2"
tiktoken = "^0.6.0"
streamlit = "^1.31.1"
pydantic = "^2.5.3"
streamlit-modal = "^0.1.2"
pymilvus = "2.3.6"
langchain-openai = "^0.0.8"
aiolimiter = "^1.1.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = {extras = ["d"], version = "^24.2.0"}
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80

[tool.isort]
profile = "black"
line_length = 80

[tool.poetry.scripts]
start = 'ai_researcher.bin.main:main'
